+++ src/components/profile/GeneralSettingsTab.js
@@ -30,8 +30,8 @@ const useStyles = makeStyles((theme) => ({
         margin: '60px 0 100px 0'
     },
     cardHeader: {
-        backgroundColor: theme.palette.type === 'light' ? '#DADADA' : '#121212',
-        color: theme.palette.type === 'light' ? '#333333' : '#DADADA',
+        backgroundColor: theme.palette.primary.main,
+        color: theme.palette.primary.contrastText,
         '& .MuiToolbar-root': {
             display: 'flex',
             justifyContent: 'space-between'